# Conflicts:
#	MediaShop.BusinessLogic.Tests/AdminTests/UserServiceTest.cs
#	MediaShop.BusinessLogic.Tests/MediaShop.BusinessLogic.Tests.csproj
#	MediaShop.BusinessLogic/MediaShop.BusinessLogic.csproj
#	MediaShop.BusinessLogic/Services/UserService.cs
#	MediaShop.Common/Exceptions/ExistingLoginException.cs
#	MediaShop.Common/MediaShop.Common.csproj
#	MediaShop.Common/Models/User/Account.cs
#	MediaShop.Common/Models/User/AccountSettings.cs

+++ MediaShop.Common/Models/Entity.cs
@@ -15,7 +15,7 @@ namespace MediaShop.Common.Models
         /// Gets or sets the identifier.
         /// </summary>
         /// <value>The identifier.</value>
-        public int Id { get; set; }
+        public long Id { get; set; }
 
         /// <summary>
         /// Gets or sets the created date.
@@ -27,7 +27,7 @@ namespace MediaShop.Common.Models
         /// Gets or sets the creator identifier.
         /// </summary>
         /// <value>The creator identifier.</value>
-        public int CreatorId { get; set; }
+        public long CreatorId { get; set; }
 
         /// <summary>
         /// Gets or sets the modified date.
@@ -39,6 +39,6 @@ namespace MediaShop.Common.Models
         /// Gets or sets the modifier identifier.
         /// </summary>
         /// <value>The modifier identifier.</value>
-        public int? ModifierId { get; set; }
+        public long? ModifierId { get; set; }
     }
 }
