+++ MediaShop.Common/Models/User/AccountSettings.cs
@@ -2,6 +2,8 @@
 // Copyright (c) MediaShop. All rights reserved.
 // </copyright>
 
+using System;
+
 namespace MediaShop.Common.Models.User
 {
     /// <summary>
@@ -37,6 +39,6 @@ namespace MediaShop.Common.Models.User
         /// <summary>
         /// id from table Account
         /// </summary>
-        public int AccountId { get; set; }
+        public long AccountId { get; set; }
     }
 }
